Physx working with changing scene
Scenes is now releasing and deleting pointers. Scenes can now change and physX works

+++ Engine/source/Project/SceneSystem/TempScenes/TempSceneOne.cpp
@@ -12,10 +12,12 @@
 
 namespace Project
 {
+    ErrorLogger log;
     TempSceneOne::TempSceneOne(CDirectX11SceneManager* sceneManager, IRenderer* renderer, int sceneIndex, CVector3 ambientColour, float specularPower, ColourRGBA backgroundColour, bool vsyncOn)
     {
         m_Renderer = renderer;
         m_SceneIndex = sceneIndex;
+
 
         m_AmbientColour = ambientColour;
         m_SpecularPower = specularPower;
@@ -23,97 +25,74 @@ namespace Project
         m_VsyncOn = vsyncOn;
 
         m_sceneManager = sceneManager;
-
-
-        /////////////////////////////////
-        // Physx set up temp (Physx 4.1 documentation startup and shutdown)
-        m_Foundation = PxCreateFoundation(PX_PHYSICS_VERSION, gDefaultAllocatorCallback, gDefaultErrorCallback);
-         if (!m_Foundation)
-            m_Log.ErrorMessage(renderer->GetWindowsProperties(), "PxCreateFoundation Failed");
-
-         m_Pvd = physx::PxCreatePvd(*m_Foundation); 
-         physx::PxPvdTransport* transport = physx::PxDefaultPvdSocketTransportCreate("127.0.0.1", 5424, 10);
-         m_Pvd->connect(*transport, physx::PxPvdInstrumentationFlag::eALL);
-
-         physx::PxTolerancesScale scale = physx::PxTolerancesScale();
-         scale.length = 0.1f;
-         //scale.speed = 0.1f;
-         m_Physics = PxCreatePhysics(PX_PHYSICS_VERSION, *m_Foundation, scale, true, m_Pvd);
-         if(!m_Physics)
-             m_Log.ErrorMessage(renderer->GetWindowsProperties(), "PxCreatePhysics Failed");
-         
-        
-        if (!PxInitExtensions(*m_Physics, m_Pvd))
-            m_Log.ErrorMessage(renderer->GetWindowsProperties(), "PxInitExtensions Failed");
+        m_EnablePhysics = false;
 
         
         
-        physx::PxCookingParams params(m_Physics->getTolerancesScale());
-        params.meshWeldTolerance = 0.001f; // Physx sample default
-        params.meshPreprocessParams = physx::PxMeshPreprocessingFlags(physx::PxMeshPreprocessingFlag::eWELD_VERTICES);
-        params.buildGPUData = true; //Enable GRB data being produced in cooking.
+      
+    }
 
-        m_Cooking = PxCreateCooking(PX_PHYSICS_VERSION, *m_Foundation, params);
-        if (!m_Cooking)
-            m_Log.ErrorMessage(renderer->GetWindowsProperties(), "PxCreateCooking Failed");
-
-
-        physx::PxSceneDesc desc(scale);
-        desc.gravity = physx::PxVec3(0.0f, -9.81f, 0.0f);
-     
-
-        if (!desc.cpuDispatcher)
-        {
-            physx::PxU32 mNbThreads = 1;
-            m_CpuDispatcher = physx::PxDefaultCpuDispatcherCreate(mNbThreads); 
-            if (!m_CpuDispatcher)
-                m_Log.ErrorMessage(renderer->GetWindowsProperties(), "CpuDispatcher Failed");
-            desc.cpuDispatcher = m_CpuDispatcher;
-        }
+    TempSceneOne::TempSceneOne(CDirectX11SceneManager* sceneManager, IRenderer* renderer, bool enablePhysics, int sceneIndex, CVector3 ambientColour, float specularPower, ColourRGBA backgroundColour, bool vsyncOn)
+    {
+        m_Renderer = renderer;
+        m_SceneIndex = sceneIndex;
+        m_EnablePhysics = enablePhysics;
         
-        if (!desc.filterShader)
-            desc.filterShader = getSampleFilterShader();
 
-        //customizeSceneDesc(desc);
-        m_Scene = m_Physics->createScene(desc);
-        if (!m_Scene)
-            m_Log.ErrorMessage(renderer->GetWindowsProperties(), "Scene Failed");
+        m_AmbientColour = ambientColour;
+        m_SpecularPower = specularPower;
+        m_backgroundColour = backgroundColour;
+        m_VsyncOn = vsyncOn;
+
+        m_sceneManager = sceneManager;
     }
 
     bool TempSceneOne::InitGeometry()
     {
+        
         m_EntityManager = new EntityManager(m_Renderer);
+        
         m_LightEntityManager = new EntityManager(m_Renderer);
+        
         m_TestManager = new EntityManager(m_Renderer);
-
+        
         CParseLevel LevelParser(m_TestManager);
-
+        
         LevelParser.ParseFile("test.xml");
-
+        
         std::string path = "media/";
-
+        
         m_EntityManager->CreateModelEntity("Cube", path + "Cube.x");
         m_EntityManager->CreateModelEntity("Cube2", path + "Cube.x", path + "brick1.jpg");
         m_EntityManager->CreateModelEntity("Ground", path + "Ground.x", path + "GrassDiffuseSpecular.dds");
-
+        
         m_EntityManager->CreateModelEntity("Crate", path + "CargoContainer.x", path + "CargoA.dds");
         m_LightEntityManager->CreateLightEntity("LightOne");
 
-        //m_Material = m_Physics->createMaterial(.5f, .5f, .1f);
-        m_Material = m_Physics->createMaterial(0, 0, 0);
 
-        m_BoxActor = m_Physics->createRigidDynamic(physx::PxTransform({ 0.0f, 40.0f, 0.0f }));
-        //m_BoxActor->setActorFlags(physx::PxActorFlag::eDISABLE_SIMULATION);
-        
+        if (m_EnablePhysics)
+        {
+             m_PhysicsSystem = NewPhysics(m_sceneManager->GetWindowsProperties().PhysicsType);
 
-        m_BoxShape = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor, physx::PxBoxGeometry(5, 5, 5), *m_Material);
-        
-        m_BoxActor2 = m_Physics->createRigidStatic({ 0.0f, 10.0f, 0.0f });
-        m_BoxShape2 = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor2, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+            if (!m_PhysicsSystem->InitPhysics())
+                m_Log.ErrorMessage(m_Renderer->GetWindowsProperties(), "Failed to Initialise Physics");
 
-        
-        m_Scene->addActor(*m_BoxActor);
-        m_Scene->addActor(*m_BoxActor2);
+            //m_Material = m_Physics->createMaterial(.5f, .5f, .1f);
+            m_Material = m_PhysicsSystem->GetPhysics()->createMaterial(0, 0, 0);
+
+            m_BoxActor = m_PhysicsSystem->GetPhysics()->createRigidDynamic(physx::PxTransform({ 0.0f, 40.0f, 0.0f }));
+            //m_BoxActor->setActorFlags(physx::PxActorFlag::eDISABLE_SIMULATION);
+
+
+            m_BoxShape = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+
+            m_BoxActor2 = m_PhysicsSystem->GetPhysics()->createRigidStatic({ 0.0f, 10.0f, 0.0f });
+            m_BoxShape2 = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor2, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+
+
+            m_PhysicsSystem->GetScene()->addActor(*m_BoxActor);
+            m_PhysicsSystem->GetScene()->addActor(*m_BoxActor2);
+        }
         
         m_SceneCamera = new Camera();
         return true;
@@ -178,32 +157,36 @@ namespace Project
     
     void TempSceneOne::UpdateScene(float frameTime)
     {
-        m_Scene->simulate(frameTime);
-        //advance(frameTime);
-        
-        if (m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
+        if (m_EnablePhysics)
         {
+            m_PhysicsSystem->GetScene()->simulate(frameTime);
+            //advance(frameTime);
             
-            TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube")->GetComponent("Transform"));
+            if (m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
+            {
+                
+                TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube")->GetComponent("Transform"));
 
-            CVector3 vect;
-            vect.x = m_BoxActor->getGlobalPose().p.x;
-            vect.y = m_BoxActor->getGlobalPose().p.y;
-            vect.z = m_BoxActor->getGlobalPose().p.z;
-            comp->SetPosition(vect);
+                CVector3 vect;
+                vect.x = m_BoxActor->getGlobalPose().p.x;
+                vect.y = m_BoxActor->getGlobalPose().p.y;
+                vect.z = m_BoxActor->getGlobalPose().p.z;
+                comp->SetPosition(vect);
 
-            renderedBox = comp->GetPosition();
-        }
+                renderedBox = comp->GetPosition();
+            }
 
-        if (m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"))
-        {
-            TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"));
+            if (m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"))
+            {
+                TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"));
 
-            CVector3 vect;
-            vect.x = m_BoxActor2->getGlobalPose().p.x;
-            vect.y = m_BoxActor2->getGlobalPose().p.y;
-            vect.z = m_BoxActor2->getGlobalPose().p.z;
-            comp->SetPosition(vect);
+                CVector3 vect;
+                vect.x = m_BoxActor2->getGlobalPose().p.x;
+                vect.y = m_BoxActor2->getGlobalPose().p.y;
+                vect.z = m_BoxActor2->getGlobalPose().p.z;
+                comp->SetPosition(vect);
+            }
+            m_PhysicsSystem->GetScene()->fetchResults(true);
         }
 
         m_EntityManager->UpdateAllEntities(frameTime);
@@ -214,33 +197,28 @@ namespace Project
 
         if (KeyHit(Key_L))
         {
-            m_sceneManager->LoadScene(1);
-            m_sceneManager->RemoveSceneAtIndex(0);
+            m_sceneManager->RemoveSceneAtIndex(1);
+            m_sceneManager->LoadScene(0);
+            
         }
 
-        m_Scene->fetchResults(true);
+      
     }
 
     void TempSceneOne::ReleaseResources()
     {
+        if (m_EntityManager         != nullptr)    m_EntityManager->DestroyAllEntities(); 
+        if (m_LightEntityManager    != nullptr)    m_LightEntityManager->DestroyAllEntities();
+        if (m_TestManager           != nullptr)    m_TestManager->DestroyAllEntities();
+        
+        SAFE_RELEASE(m_Material);
+        SAFE_RELEASE(m_BoxActor);
+        SAFE_RELEASE(m_BoxActor2);
 
-        m_EntityManager->DestroyAllEntities();
-        m_LightEntityManager->DestroyAllEntities();
-        m_TestManager->DestroyAllEntities();
+        if(m_PhysicsSystem != nullptr) m_PhysicsSystem->ShutdownPhysics();
 
-        m_Physics->release();       delete m_Physics;       m_Physics = nullptr;
-        m_Foundation->release();    m_Foundation = nullptr;
-        m_Cooking->release();       m_Cooking = nullptr;
-        m_Pvd->release();           m_Pvd = nullptr;
-        m_Scene->release();         m_Scene = nullptr;
 
-        m_BoxActor->release();
-        m_BoxActor2->release();
-        m_BoxShape->release();
-        m_Material->release();
-
-        delete m_SceneCamera;       m_SceneCamera = nullptr;
-        delete m_CpuDispatcher;     m_CpuDispatcher = nullptr;
+        if (m_SceneCamera != nullptr) { delete m_SceneCamera;  m_SceneCamera = nullptr; }
     }
 
     void TempSceneOne::onContact(const physx::PxContactPairHeader& pairHeader, const physx::PxContactPair* pairs, physx::PxU32 nbPairs)
@@ -301,13 +279,13 @@ namespace Project
 
   bool TempSceneOne::advance(physx::PxReal dt)
     {
-        mAccumulator += dt;
+       /* mAccumulator += dt;
         if (mAccumulator < mStepSize)
             return false;
 
         mAccumulator -= mStepSize;
 
-        m_Scene->simulate(mStepSize);
+        m_Scene->simulate(mStepSize);*/
         return true;
     }
 }
