+++ Engine/source/Project/SceneSystem/TempScenes/TempSceneThree.cpp
@@ -6,7 +6,7 @@
 
 namespace Project
 {
-	/*TempSceneThree::TempSceneThree(CDirectX11SceneManager* sceneManager, IRenderer* renderer, int sceneIndex, CVector3 ambientColour, float specularPower, ColourRGBA backgroundColour, bool vsyncOn)
+	TempSceneThree::TempSceneThree(CDirectX11SceneManager* sceneManager, IRenderer* renderer, int sceneIndex, CVector3 ambientColour, float specularPower, ColourRGBA backgroundColour, bool vsyncOn)
 	{
 		m_Renderer = renderer;
 		m_SceneIndex = sceneIndex;
@@ -19,15 +19,13 @@ namespace Project
 		m_sceneManager = sceneManager;
 
 		m_EnablePhysics = false;
-	}*/
+	}
 
 	TempSceneThree::TempSceneThree(CDirectX11SceneManager* sceneManager, IRenderer* renderer, bool enablePhysics, int sceneIndex, CVector3 ambientColour, float specularPower, ColourRGBA backgroundColour, bool vsyncOn)
 	{
 		m_Renderer = renderer;
 		m_SceneIndex = sceneIndex;
 		m_EnablePhysics = enablePhysics;
-		if (m_EnablePhysics)
-			m_PhysicsSystem = NewPhysics(sceneManager->GetWindowsProperties().PhysicsType);
 
 
 		m_AmbientColour = ambientColour;
@@ -41,37 +39,45 @@ namespace Project
 
 	bool TempSceneThree::InitGeometry()
 	{
-		if (m_EnablePhysics)
-			if(!m_PhysicsSystem->InitPhysics())
-				m_Log.ErrorMessage(m_Renderer->GetWindowsProperties(), "Failed to Initialise Physics");
-
+		
 		m_EntityManager = new EntityManager(m_Renderer);
-
+		
 		m_SceneCamera = new Camera();
-
+		
 		std::string path = "media/";
 		m_EntityManager->CreateModelEntity("Floor", path + "Ground.x");
 
 		m_EntityManager->CreateModelEntity("Cube", path + "Cube.x");
 		m_EntityManager->CreateModelEntity("Cube2", path + "Cube.x", path + "brick1.jpg");
+		
+		if (m_EnablePhysics)
+		{
+			m_PhysicsSystem = NewPhysics(m_sceneManager->GetWindowsProperties().PhysicsType);
 
-		m_Material = m_PhysicsSystem->GetPhysics()->createMaterial(0, 0, 0);
-		m_BoxActor = m_PhysicsSystem->GetPhysics()->createRigidDynamic(physx::PxTransform({ 0.0f, 40.0f, 0.0f }));
-		m_BoxShape = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+			if (!m_PhysicsSystem->InitPhysics())
+				m_Log.ErrorMessage(m_Renderer->GetWindowsProperties(), "Failed to Initialise Physics");
 
-		m_BoxActor2 = m_PhysicsSystem->GetPhysics()->createRigidStatic({ 0.0f, 10.0f, 0.0f });
-		m_BoxShape2 = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor2, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+			
+
+			m_Material = m_PhysicsSystem->GetPhysics()->createMaterial(0, 0, 0);
+			m_BoxActor = m_PhysicsSystem->GetPhysics()->createRigidDynamic(physx::PxTransform({ 0.0f, 40.0f, 0.0f }));
+			m_BoxShape = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor, physx::PxBoxGeometry(5, 5, 5), *m_Material);
+
+			m_BoxActor2 = m_PhysicsSystem->GetPhysics()->createRigidStatic({ 0.0f, 10.0f, 0.0f });
+			m_BoxShape2 = physx::PxRigidActorExt::createExclusiveShape(*m_BoxActor2, physx::PxBoxGeometry(5, 5, 5), *m_Material);
 
 
-		m_PhysicsSystem->GetScene()->addActor(*m_BoxActor);
-		m_PhysicsSystem->GetScene()->addActor(*m_BoxActor2);
+			m_PhysicsSystem->GetScene()->addActor(*m_BoxActor);
+			m_PhysicsSystem->GetScene()->addActor(*m_BoxActor2);
+		}
 
 		return true;
 	}
 
 	bool TempSceneThree::InitScene()
 	{
-		if (m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
+		
+		if (m_EnablePhysics && m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
 		{
 			TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube")->GetComponent("Transform"));
 
@@ -101,43 +107,65 @@ namespace Project
 
 	void TempSceneThree::UpdateScene(float frameTime)
 	{
-		m_PhysicsSystem->GetScene()->simulate(frameTime);
-
-		if (m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
+		if (m_EnablePhysics)
 		{
+			m_PhysicsSystem->GetScene()->simulate(frameTime);
 
-			TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube")->GetComponent("Transform"));
+			if (m_EntityManager->GetEntity("Cube")->GetComponent("Transform"))
+			{
 
-			CVector3 vect;
-			vect.x = m_BoxActor->getGlobalPose().p.x;
-			vect.y = m_BoxActor->getGlobalPose().p.y;
-			vect.z = m_BoxActor->getGlobalPose().p.z;
-			comp->SetPosition(vect);
-		}
+				TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube")->GetComponent("Transform"));
 
-		if (m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"))
-		{
-			TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"));
+				CVector3 vect;
+				vect.x = m_BoxActor->getGlobalPose().p.x;
+				vect.y = m_BoxActor->getGlobalPose().p.y;
+				vect.z = m_BoxActor->getGlobalPose().p.z;
+				comp->SetPosition(vect);
+			}
 
-			CVector3 vect;
-			vect.x = m_BoxActor2->getGlobalPose().p.x;
-			vect.y = m_BoxActor2->getGlobalPose().p.y;
-			vect.z = m_BoxActor2->getGlobalPose().p.z;
-			comp->SetPosition(vect);
+			if (m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"))
+			{
+				TransformComponent* comp = static_cast<TransformComponent*>(m_EntityManager->GetEntity("Cube2")->GetComponent("Transform"));
+
+				CVector3 vect;
+				vect.x = m_BoxActor2->getGlobalPose().p.x;
+				vect.y = m_BoxActor2->getGlobalPose().p.y;
+				vect.z = m_BoxActor2->getGlobalPose().p.z;
+				comp->SetPosition(vect);
+			}
+			m_PhysicsSystem->GetScene()->fetchResults(true);
 		}
 
 		m_EntityManager->UpdateAllEntities(frameTime);
 
 		m_SceneCamera->Control(frameTime);
-
-		m_PhysicsSystem->GetScene()->fetchResults(true);
+		if (KeyHit(Key_L))
+		{
+			
+			m_sceneManager->RemoveSceneAtIndex(0);
+			
+			m_sceneManager->LoadScene(1);
+			
+		}
+		
 	}
 
 	void TempSceneThree::ReleaseResources()
 	{
-		delete m_SceneCamera;     m_SceneCamera = nullptr;
+		
+		if (m_SceneCamera != nullptr) { delete m_SceneCamera;  m_SceneCamera = nullptr; }
 
-		m_EntityManager->DestroyAllEntities();
-		m_PhysicsSystem->ShutdownPhysics();
+		if (m_EntityManager != nullptr)    m_EntityManager->DestroyAllEntities();
+
+		SAFE_RELEASE(m_Material);
+
+		SAFE_RELEASE(m_BoxActor);
+
+		SAFE_RELEASE(m_BoxActor2);
+		
+		
+		if (m_PhysicsSystem != nullptr) m_PhysicsSystem->ShutdownPhysics();
+		
+		
 	}
 }
